Change pass for events components
