+++ eventually/event/models.py
@@ -119,7 +119,7 @@ class Event(AbstractModel):
             'start_at': int(self.start_at.timestamp()) if self.start_at else None,
             'created_at': int(self.created_at.timestamp()),
             'updated_at': int(self.updated_at.timestamp()),
-            'duration': self.duration.seconds if self.duration else None,
+            'duration': self.duration.total_seconds() if self.duration else None,
             'longitude': float(self.longitude) if self.longitude else None,
             'latitude': float(self.latitude) if self.latitude else None,
             'budget': self.budget,
