+++ eventually/static/src/containers/event/CreateEvent.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import { GetTeamsListService, PostEventService } from './CreateEventService';
+import { GetTeamsListService, PostEventService } from './EventService';
 import Dialog from 'material-ui/Dialog';
 import FlatButton from 'material-ui/FlatButton';
 import RaisedButton from 'material-ui/RaisedButton';
